+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ==========
 
 ### Fixed
+- Fixed `GetExports` only processing the first page of DescribeStacks results, causing exports from accounts with more than 100 stacks to be omitted
 - Fixed `GetResources` only processing the first page of DescribeStacks results, causing resources from accounts with more than 100 stacks to be omitted
 
 1.12.2 / 2026-01-19